Refactor file paths in seed.ts and database configuration files
